+++ renren-admin/src/main/java/io/renren/modules/sys/service/SysUserService.java
@@ -98,7 +98,7 @@ public interface SysUserService extends IService<SysUserEntity> {
 
     Long selectCountByUserName(PageData pageData) throws Exception;
 
-    Integer queryByUserName(PageData pageData) throws Exception;
+    Long queryByUserName(PageData pageData) throws Exception;
 
     PageData selectEmailAndPhoneByUserName(PageData pageData) throws Exception;
